requirements.txt
altgraph==0.17.4
asgiref==3.8.1
attrs==23.1.0
autopep8==1.6.0
certifi==2023.7.22
cffi==1.15.1
channels==4.1.0
chardet==5.2.0
charset-normalizer==3.3.2
contourpy==1.2.1
coreapi==2.3.3
coreschema==0.0.4
crispy-bootstrap5==2024.2
cryptography==42.0.8
cycler==0.12.1
defusedxml==0.8.0rc2
dj-database-url==0.5.0
Django==5.0.6
django-allauth==0.63.3
django-appconf==1.0.6
django-cors-headers==4.2.0
django-crispy-forms==2.2
django-environ==0.8.1
django-filter==23.3
django-imagekit==5.0.0
django-phonenumber-field==7.1.0
django-taggit==4.0.0
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
et_xmlfile==2.0.0
exceptiongroup==1.1.3
ffmpeg==1.4
fonttools==4.50.0
gunicorn==20.1.0
h11==0.14.0
idna==3.4
imageio==2.34.0
imageio-ffmpeg==0.4.9
itypes==1.2.0
Jinja2==3.1.6
jwt==1.3.1
keyboard==0.13.5
kiwisolver==1.4.5
Markdown==3.4.4
MarkupSafe==3.0.2
matplotlib==3.8.4
MouseInfo==0.1.3
mysqlclient==2.2.4
numpy==1.26.4
oauthlib==3.2.2
openpyxl==3.1.5
outcome==1.2.0
packaging==23.2
pandas==2.2.3
pefile==2023.2.7
phonenumbers==8.13.21
pilkit==3.0
Pillow==10.4.0
psutil==5.9.8
psycopg2==2.9.7
PyAutoGUI==0.9.54
pycodestyle==2.8.0
pycparser==2.21
pygame==2.6.1
PyGetWindow==0.0.9
pyinstaller==6.2.0
pyinstaller-hooks-contrib==2023.10
PyJWT==2.8.0
PyMsgBox==1.0.9
PyOpenGL==3.1.9
pyparsing==3.1.2
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.1.0
python3-openid==3.2.0
pytweening==1.2.0
pytz==2021.3
pywin32-ctypes==0.2.2
reportlab==4.2.2
requests==2.32.3
requests-oauthlib==2.0.0
scipy==1.15.0
selenium==4.12.0
six==1.16.0
sniffio==1.3.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sortedcontainers==2.4.0
sqlparse==0.5.0
toml==0.10.2
trimesh==4.6.8
trio==0.22.2
trio-websocket==0.10.4
typing_extensions==4.12.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.0.4
vobject==0.9.6.1
whitenoise==5.3.0
wsproto==1.2.0
